Feat: 구조 재설정
