+++ src/main/java/com/knu/daeguhackathon/building/controller/BuildingController.java
@@ -1,0 +1,9 @@
+package com.knu.daeguhackathon.building.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+
+@Controller
+@RequiredArgsConstructor
+public class BuildingController {
+}
