+++ src/main/java/com/knu/daeguhackathon/building/service/BuildingService.java
@@ -1,0 +1,9 @@
+package com.knu.daeguhackathon.building.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class BuildingService {
+}
