--- src/main/java/com/knu/daeguhackathon/entity/Building.java
@@ -1,33 +1,0 @@
-package com.knu.daeguhackathon.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToOne;
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Building {
-
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Id
-    private Long id;
-    @NotNull
-    private String buildingName;
-    @NotNull
-    private Double latitude;
-    @NotNull
-    private Double longitude;
-    @NotNull
-    @OneToOne
-    private Subject subject;
-}
