+++ src/main/java/com/knu/daeguhackathon/kakaotoken/KakaoToken.java
@@ -1,5 +1,6 @@
-package com.knu.daeguhackathon.entity;
+package com.knu.daeguhackathon.kakaotoken;
 
+import com.knu.daeguhackathon.member.Member;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -8,13 +9,11 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.OneToOne;
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDateTime;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+
+import lombok.*;
 
 @Entity
-@Data
+@Getter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
