+++ src/main/java/com/knu/daeguhackathon/member/Member.java
@@ -1,5 +1,8 @@
-package com.knu.daeguhackathon.entity;
+package com.knu.daeguhackathon.member;
 
+import com.knu.daeguhackathon.searchlog.SearchLog;
+import com.knu.daeguhackathon.kakaotoken.KakaoToken;
+import com.knu.daeguhackathon.subject.Subject;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -12,13 +15,11 @@ import jakarta.validation.constraints.NotNull;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+
+import lombok.*;
 
 @Entity
-@Data
+@Getter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
@@ -43,5 +44,5 @@ public class Member {
     @OneToMany(mappedBy = "member", cascade = CascadeType.REMOVE)
     private List<Subject> subjects = new ArrayList<>();
     @OneToMany(mappedBy = "member", cascade = CascadeType.REMOVE)
-    private List<searchLog> searchLogs = new ArrayList<>();
+    private List<SearchLog> SearchLogs = new ArrayList<>();
 }
