+++ src/main/java/com/knu/daeguhackathon/member/controller/MemberController.java
@@ -1,0 +1,9 @@
+package com.knu.daeguhackathon.member.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+
+@Controller
+@RequiredArgsConstructor
+public class MemberController {
+}
