+++ src/main/java/com/knu/daeguhackathon/member/service/MemberService.java
@@ -1,0 +1,9 @@
+package com.knu.daeguhackathon.member.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class MemberService {
+}
