+++ src/main/java/com/knu/daeguhackathon/searchlog/SearchLog.java
@@ -1,5 +1,6 @@
-package com.knu.daeguhackathon.entity;
+package com.knu.daeguhackathon.searchlog;
 
+import com.knu.daeguhackathon.member.Member;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -8,17 +9,15 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDateTime;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+
+import lombok.*;
 
 @Entity
-@Data
+@Getter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class searchLog {
+public class SearchLog {
 
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
