+++ src/main/java/com/knu/daeguhackathon/subject/Subject.java
@@ -1,0 +1,35 @@
+package com.knu.daeguhackathon.subject;
+
+import com.knu.daeguhackathon.building.Building;
+import com.knu.daeguhackathon.member.Member;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToOne;
+import jakarta.validation.constraints.NotNull;
+import lombok.*;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Subject {
+
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    private Long id;
+    @NotNull
+    private String courseName;
+    @NotNull
+    @ManyToOne
+    @JoinColumn(name = "member_id")
+    private Member member;
+    @NotNull
+    @OneToOne
+    @JoinColumn(name = "building_id")
+    private Building building;
+}
