+++ src/main/java/com/knu/daeguhackathon/subject/controller/SubjectController.java
@@ -1,0 +1,9 @@
+package com.knu.daeguhackathon.subject.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+
+@Controller
+@RequiredArgsConstructor
+public class SubjectController {
+}
