+++ src/main/java/com/knu/daeguhackathon/subject/service/SubjectService.java
@@ -1,0 +1,9 @@
+package com.knu.daeguhackathon.subject.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class SubjectService {
+}
